Fix Double-close on a filehandle (RT93221)
https://rt.cpan.org/Public/Bug/Display.html?id=93221

The problem is that _magic() closes the fd when it finds a result
for lower priority rules. mimetype() then tries to close it again.

Since we take a completely different route in mimetype() if
'ref file' is true, we can close the fd safely whenever the
low-priority rules return one.